+++ geotrek/common/management/commands/test_managers_emails.py
@@ -1,4 +1,3 @@
-from django.conf import settings
 from django.core.mail import mail_managers
 from django.core.management.base import BaseCommand
 
@@ -6,9 +5,8 @@ from django.core.management.base import BaseCommand
 class Command(BaseCommand):
     help = "Test if email settings are OK by sending mail to site managers"
 
-    def execute(self, *args, **options):
-
-        subject = u'[%s] Test email for managers' % settings.TITLE
+    def handle(self, *args, **options):
+        subject = u'Test email for managers'
         message = u'If you receive this email, configuration is OK !'
 
         mail_managers(subject, message, fail_silently=False)