+++ docs/changelog.rst
@@ -2,13 +2,33 @@
 CHANGELOG
 =========
 
+2.21.0 (2018-09-04)
+-------------------
+
+**New features**
+
+* Path deletion warning page now shows linked topologies
+* Allow to add Dailymotion video attachments
+* Add a command to unset structure in lists of choices and group choices
+  with the same name
+
+**Bug fixes**
+
+* Fix Sync_rando View
+* Fix loaddem
+* Fix creation of Youtube/Soundcloud attachments
+* Fix cancellation when editing geometries
+* Show which structure choices are related to
+* Add confort and stake filters to path list
+* Fix sync of touristic contents for mobile app
+
+
 2.20.1 (2018-07-16)
 -------------------
 
 **Bug fixes**
 
 * Fix Completed Filter in Touristic Event
-
 
 2.20.0 (2018-06-27)
 -------------------
@@ -18,8 +38,8 @@ CHANGELOG
 * Allow to share glossaries between structure (just keep structure field empty)
 * Allow to import infrastructures, not only signages
 * Allow to split treks by itinerancy
-* Path deletion does not delete the related point topologies anymore.
-  Instead point topologies are linked to the nearest remaining path.
+* Path deletion does not delete the related point topologies anymore. Instead point topologies are linked to the nearest
+  remaining path.
 * Keep inode and mtime when synced file don't change
 
 **Minor changes**
@@ -30,7 +50,7 @@ CHANGELOG
 
 * Fix virtualenv install
 * Upgrade celery to 4.1.1
-* Fix the bug which remove a path when we merge 2 paths with a 3rd path on the point of merge (ref  #1747).
+* Fix the bug which remove a path when we merge 2 paths with a 3rd path on the point of merge. (ref  #1747)
 
 
 2.19.1 (2018-05-28)